requirements_dm21.txt
absl-py==1.4.0
astunparse==1.6.3
attrs==22.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
#cupy-cuda12x==13.4.1
#cutensor-cu12==2.2.0
exceptiongroup==1.2.2
fastrlock==0.8.3
flatbuffers==24.3.25
gast==0.4.0
geometric==1.1
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
#gpu4pyscf-cuda12x==1.4.0
#gpu4pyscf-libxc-cuda12x==0.5
grpcio==1.67.1
h5py==3.8.0
idna==3.10
iniconfig==2.0.0
jsonpickle==4.0.5
#keras==2.11.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
narwhals==1.34.0
networkx==3.4.2
numpy==1.24.1
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.14.1
packaging==24.1
plotly==6.0.1
pluggy==0.13.1
progress-table==3.1.2
protobuf==3.19.6
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.1
pyscf==2.1.1
pyscf-dispersion==1.3.0
pytest==6.2.4
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scipy==1.9.3
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
#tf_keras==2.19.0
toml==0.10.2
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.16.0
pandas
